+++ CHANGELOG.md
@@ -38,6 +38,8 @@ with the `YY.MINOR.MICRO` scheme.
 - Node construction is now deferred until actually needed, avoiding unnecessary 
   work for cached queries on any instance execution
 - Fixed a memory leak in SQL API of plugin SDK
+- `picodata status` no longer panics when `stdout`, `stderr`,
+  or both are redirected to a broken pipe.
 
 ## [25.5.1] - 2025-12-19